base: Remove fullscreen recents [1/2]
Currently we have a way to set fullscreen/statusbar only/navbar only for any
app we have on our device (system or user). This includes having the ability
to set these options to SystemUI which covers this mod.

It was fun while it lasted but is much easier to set a boolean to false from true
than to add code to the fwb and settings. Work smarter not harder!

Change-Id: I5c1b087c2a334f80d6b38ec03459336e1539dc17